requirements.txt
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
captcha==0.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
pdfminer.six==20201018
pdfminer3k==1.3.4
Pillow==8.1.0
ply==3.11
psutil==5.8.0
pycparser==2.20
pycryptodome==3.10.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.2
SQLAlchemy==1.3.23
sqlparse==0.4.1
urllib3==1.26.3
Werkzeug==1.0.1
